Add support for Imenu
Imenu is an Emacs features that allows a package to define the outline
of the buffer's content to let user navigate quickly in it.

This commit brings Imenu support to `hoa-pp-mode` by creating two
sections, "Tokens" and "Rules", that lets the user go quickly to the
token / rule definition of its choice.